Fix lessThan operator not applying column case
