+++ packages/twenty-server/src/engine/api/graphql/graphql-query-runner/utils/compute-where-condition-parts.ts
@@ -1,0 +1,88 @@
+import { ObjectLiteral } from 'typeorm';
+
+import {
+  GraphqlQueryRunnerException,
+  GraphqlQueryRunnerExceptionCode,
+} from 'src/engine/api/graphql/graphql-query-runner/errors/graphql-query-runner.exception';
+
+type WhereConditionParts = {
+  sql: string;
+  params: ObjectLiteral;
+};
+
+export const computeWhereConditionParts = (
+  operator: string,
+  objectNameSingular: string,
+  key: string,
+  value: any,
+): WhereConditionParts => {
+  const uuid = Math.random().toString(36).slice(2, 7);
+
+  switch (operator) {
+    case 'eq':
+      return {
+        sql: `"${objectNameSingular}"."${key}" = :${key}${uuid}`,
+        params: { [`${key}${uuid}`]: value },
+      };
+    case 'neq':
+      return {
+        sql: `"${objectNameSingular}"."${key}" != :${key}${uuid}`,
+        params: { [`${key}${uuid}`]: value },
+      };
+    case 'gt':
+      return {
+        sql: `"${objectNameSingular}"."${key}" > :${key}${uuid}`,
+        params: { [`${key}${uuid}`]: value },
+      };
+    case 'gte':
+      return {
+        sql: `"${objectNameSingular}"."${key}" >= :${key}${uuid}`,
+        params: { [`${key}${uuid}`]: value },
+      };
+    case 'lt':
+      return {
+        sql: `"${objectNameSingular}"."${key}" < :${key}${uuid}`,
+        params: { [`${key}${uuid}`]: value },
+      };
+    case 'lte':
+      return {
+        sql: `"${objectNameSingular}"."${key}" <= :${key}${uuid}`,
+        params: { [`${key}${uuid}`]: value },
+      };
+    case 'in':
+      return {
+        sql: `"${objectNameSingular}"."${key}" IN (:...${key}${uuid})`,
+        params: { [`${key}${uuid}`]: value },
+      };
+    case 'is':
+      return {
+        sql: `"${objectNameSingular}"."${key}" IS ${value === 'NULL' ? 'NULL' : 'NOT NULL'}`,
+        params: {},
+      };
+    case 'like':
+      return {
+        sql: `"${objectNameSingular}"."${key}" LIKE :${key}${uuid}`,
+        params: { [`${key}${uuid}`]: `${value}` },
+      };
+    case 'ilike':
+      return {
+        sql: `"${objectNameSingular}"."${key}" ILIKE :${key}${uuid}`,
+        params: { [`${key}${uuid}`]: `${value}` },
+      };
+    case 'startsWith':
+      return {
+        sql: `"${objectNameSingular}"."${key}" LIKE :${key}${uuid}`,
+        params: { [`${key}${uuid}`]: `${value}` },
+      };
+    case 'endsWith':
+      return {
+        sql: `"${objectNameSingular}"."${key}" LIKE :${key}${uuid}`,
+        params: { [`${key}${uuid}`]: `${value}` },
+      };
+    default:
+      throw new GraphqlQueryRunnerException(
+        `Operator "${operator}" is not supported`,
+        GraphqlQueryRunnerExceptionCode.UNSUPPORTED_OPERATOR,
+      );
+  }
+};
